Спринт 9. Дз № 3. Добавил поле для ввода текста (EditText и TextWatcher) на экране Поиска.
