+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.MaterialComponents.DayNight.NoActionBar"
         tools:targetApi="31">
         <activity
-            android:name=".Search"
+            android:name=".SearchActivity"
             android:exported="false"
             android:theme="@style/Theme.PlaylistMaker.Search">
             <meta-data
