--- app/src/main/java/com/practicum/playlistmaker/Search.kt
@@ -1,20 +1,0 @@
-package com.practicum.playlistmaker
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.widget.ImageView
-
-class Search : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_search)
-
-        val buttonGoBack = findViewById<ImageView>(R.id.button_go_back)
-
-        buttonGoBack.setOnClickListener {
-            val displayMainMenu = Intent(this, MainActivity::class.java)
-            startActivity(displayMainMenu)
-        }
-    }
-}
