+++ app/src/main/java/com/practicum/playlistmaker/SearchActivity.kt
@@ -1,0 +1,55 @@
+package com.practicum.playlistmaker
+
+import android.content.Context
+import android.content.Intent
+import android.os.Bundle
+import android.text.Editable
+import android.text.TextWatcher
+import android.view.View
+import android.view.inputmethod.InputMethodManager
+import android.widget.EditText
+import android.widget.ImageView
+import androidx.appcompat.app.AppCompatActivity
+
+class SearchActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_search)
+
+        val buttonGoBack = findViewById<ImageView>(R.id.button_go_back)
+
+        buttonGoBack.setOnClickListener {
+            val displayMainMenu = Intent(this, MainActivity::class.java)
+            startActivity(displayMainMenu)
+        }
+
+        val inputEditText = findViewById<EditText>(R.id.inputEditText)
+        val clearButton = findViewById<ImageView>(R.id.clearIcon)
+
+        clearButton.setOnClickListener {
+            inputEditText.setText("")
+            val inputMethodManager = getSystemService(Context.INPUT_METHOD_SERVICE) as? InputMethodManager
+            inputMethodManager?.hideSoftInputFromWindow(currentFocus!!.windowToken, 0)
+        }
+
+        val simpleTextWatcher = object : TextWatcher {
+            override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {
+                // пока пусто
+            }
+            override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {
+                clearButton.visibility = clearButtonVisibility(s)
+            }
+            override fun afterTextChanged(s: Editable?) {
+                // пока пусто
+            }
+        }
+        inputEditText.addTextChangedListener(simpleTextWatcher)
+    }
+    private fun clearButtonVisibility(s: CharSequence?): Int {
+        return if (s.isNullOrEmpty()) {
+            View.GONE
+        } else {
+            View.VISIBLE
+        }
+    }
+}
